+++ apps/video/CHANGELOG.md
@@ -1,5 +1,19 @@
 # web
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/video@0.1.5
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/video@0.1.3588
+
 ## 0.1.1
 
 ### Patch Changes